refactor: :truck: move location to common
