+++ src/auth/dto/user.dto.ts
@@ -1,5 +1,5 @@
 import EventDetails from '../../event/entity/event.details.entity';
-import Location from '../../location/entity/location.entity';
+import Location from '../../common/entity/location.entity';
 import Preference from '../../common/entity/preference.entity';
 
 class UserDTO {
