+++ src/auth/entity/user.entity.ts
@@ -13,7 +13,7 @@ import {
 } from 'typeorm';
 import Preference from '../../common/entity/preference.entity';
 import Event from '../../event/entity/event.entity';
-import Location from '../../location/entity/location.entity';
+import Location from '../../common/entity/location.entity';
 import Image from '../../image/entity/image.entity';
 
 @Entity()
