+++ src/event/entity/event.entity.ts
@@ -17,7 +17,7 @@ import Status from '../../common/entity/status.entity';
 
 import User from '../../auth/entity/user.entity';
 import Image from '../../image/entity/image.entity';
-import Location from '../../location/entity/location.entity';
+import Location from '../../common/entity/location.entity';
 
 @Entity()
 class Event {
